requirements.txt
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
cryptography==46.0.5
idna==3.11
psycopg2-binary==2.9.11
pycparser==3.0
pyTelegramBotAPI==4.32.0
python-dotenv==1.2.2
requests==2.32.5
schedule==1.2.2
urllib3==2.6.3
